updating athena refresh function to use new athena client class
* introduces new AthenaRefresher class that is used by the athena refresh function
* Updating sqs client class and moving to a new clients.py file
* Adding unit tests and updating all old unit tests
